+++ ls/CHANGELOG.md
@@ -13,6 +13,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Fixed
+
+-  Make sure to strip quotes around command line arguments.
+
+
 [0.12.1] - 2026-05-22
 ---------------------